fix ordering error due to actor usage
mapping a stream async using an actor method wrapped in Future/Task
caused events to be processed out of order (since the call to `apply`
wasn't guaranteed to happen in the order of events). Instead, use a DIY
"actor" with an explicit queue that events can be scheduled on (similar
to old inline approach).